Refactor : ErrorFallback 적용 & api, db 위치 변경 (#61)
* Refactor : 파일 위치 변경, 주석 삭제, 경로 변경

* Refactor : API 수정

* Refactor : Tab 공통 컴포넌트

* Refactor : Storage 컴포넌트 분리

* Refactor : play에 object fit

* Refactor : 컴포넌트, 훅 기반 FollowInfo 수정

* Refactor : Profile 컴포넌트 분리 및 적용

* Refactor : QueryKey Constant 수정

* Refactor : 안 쓰는 api 주석처리, isError, isLoading 추가

* Refactor : 위치 변경

* Refactor : FollowInfo scroll 적용

* Refactor : ErrorFallback 컴포넌트 제작 및 적용

* Refactor : storage ErrorFallback 컴포넌트 속 Title 제거

* Refactor : StorageProfileCard 객체 전달

* Refactor : play, detail, modify에  ErrorFallback 적용

* Refactor : api & db 위치 변경

* Refactor : types 폴더 이동

* Refactor : queries 폴더 삭제

* Refactor : svg import 수정
