+++ src/pages/homeAndSearch/hooks/useLikedPlaylistIds.ts
@@ -1,5 +1,5 @@
 import { useQuery } from "@tanstack/react-query";
-import { getLikedPlaylistIds } from "@/api/likeActive";
+import { getLikedPlaylistIds } from "@/shared/api/likeActive";
 
 const useLikedPlaylistIds = (randomId?: number) => {
   return useQuery({
