+++ src/pages/profile/hooks/useUploadDeleteProfileImage.ts
@@ -1,6 +1,6 @@
-import { supabase } from "@/lib/supabase";
+import { supabase } from "@/shared/lib/supabase";
 import { useMutation } from "@tanstack/react-query";
-import { updateUser } from "@/db/users";
+import { updateUser } from "@/shared/api/users";
 import useUser from "@/shared/hooks/useUser";
 
 interface UploadDeleteArgs {
